Fix multiline editing
In an attempt to solve the chat input box's horrible vertical alignment
on Firefox the line-height and height was set manually. This lead to the
input not expanding with multi-line text input.

Now this change has been reverted back to the borked state because it
might look awful, but editing single-line only fucking blows.