Add simple FF userstyle manager
Using userChrome.css and userContent.css it is possible to do what most
userstyle manager addons do. The only disadvantage is that it cannot be
done live, you have to restart the browser.

The `./update.sh` script is responsible for assembling the two `.css`
files. The `userChrome.css` takes everything inside the `userChrome`
folder and concatenates it. It also adds the correct header for the FF
UI domain.

The `userContent.css` works in a similar way, but here the file names
are used to set the URL the style applies to.