Fix rustdoc outside of actual documentation
Because rustdoc is not only documentation, the rest of the website was
looking very odd. This did not require much changes to make the whole
website look uniform, so including it in this theme seems like a
reasonable idea. The website should now be completely themed expect for
some minor details.

This fixes #21.