Rework button colors
By default discord uses blue buttons. Having a big part of your UI in
bright blue doesn't seem like that great of an idea to me. However these
buttons are reused everywhere so setting it to one of our two background
colors would lead to inconsistencies.

To make sure the buttons are always clearly visible and it is easy to
read them and realize when you hover above them, the secondary
foreground has been used as background and the background color has been
used as foreground color.