Deprecate twitter theme
The twitter theme has been moved to the new `unmaintained` folder since
it has not been updated in a long time. The readme was updated to
explain this and the outdated image has been removed.

The 'Firefox' folder has also been renamed to 'firefox' to be
consistent with the new 'unmaintained' folder.