Use Firefox Color extension for coloring
To ensure the userChrome has to be edited as little as possible, the
majority of color changes are now made through the Firefox Color
extension, which does not rely on CSS name consistency.

The URL color in the URL search history dropdown still has to be set
manually since it cannot be changed with Firefox Color.